move media class to own file
